updated test and moved common data to utils
