+++ care/facility/management/commands/load_icd11_diagnoses_data.py
@@ -1,114 +1,7 @@
-import json
-
 from django.core.management import BaseCommand, CommandError
 
 from care.facility.models.icd11_diagnosis import ICD11Diagnosis
-
-
-def fetch_data():
-    with open("data/icd11.json", "r") as json_file:
-        return json.load(json_file)
-
-
-ICD11_ID_SUFFIX_TO_INT = {
-    "mms": 1,
-    "other": 2,
-    "unspecified": 3,
-}
-
-
-def icd11_id_to_int(icd11_id):
-    """
-    Maps ICD11 ID to an integer.
-
-    Eg:
-    - http://id.who.int/icd/entity/594985340 -> 594985340
-    - http://id.who.int/icd/entity/594985340/mms -> 5949853400001
-    - http://id.who.int/icd/entity/594985340/mms/unspecified -> 5949853400003
-    """
-    entity_id = icd11_id.replace("http://id.who.int/icd/entity/", "")
-    if entity_id.isnumeric():
-        return int(entity_id)
-    segments = entity_id.split("/")
-    return int(segments[0]) * 1e3 + ICD11_ID_SUFFIX_TO_INT[segments[-1]]
-
-
-CLASS_KIND_DB_KEYS = {
-    "chapter": "meta_chapter",
-    "block": "meta_root_block",
-    "category": "meta_root_category",
-}
-
-ICD11_GROUP_LABEL_PRETTY = {
-    "01 Certain infectious or parasitic diseases": "Infection and parasites ",
-    "02 Neoplasms": "Neoplasms",
-    "03 Diseases of the blood or blood-forming organs": "Hematology",
-    "04 Diseases of the immune system": "Immune System",
-    "05 Endocrine, nutritional or metabolic diseases": "Endocrine, nutritional or metabolism",
-    "06 Mental, behavioural or neurodevelopmental disorders": "Mental, behavioural or neurological",
-    "07 Sleep-wake disorders": "Sleep",
-    "08 Diseases of the nervous system": "Nervous System",
-    "09 Diseases of the visual system": "Ophthalmology",
-    "10 Diseases of the ear or mastoid process": "ENT",
-    "11 Diseases of the circulatory system": "Circulatory System",
-    "12 Diseases of the respiratory system": "Respiratory System",
-    "13 Diseases of the digestive system": "Digestive System",
-    "14 Diseases of the skin": "Skin",
-    "15 Diseases of the musculoskeletal system or connective tissue": "Musculoskeletal System",
-    "16 Diseases of the genitourinary system": "Genitourinary System",
-    "17 Conditions related to sexual health": "Sexual Health",
-    "18 Pregnancy, childbirth or the puerperium": "OBGYN",
-    "19 Certain conditions originating in the perinatal period": "Neonatology",
-    "20 Developmental anomalies": "Developmental Anomalies",
-    "21 Symptoms, signs or clinical findings, not elsewhere classified": "Others",
-    "22 Injury, poisoning or certain other consequences of external causes": "Injury, Poisoning",
-    "23 External causes of morbidity or mortality": "External Causes of Injury",
-    "24 Factors influencing health status or contact with health services": None,
-    "25 Codes for special purposes": "Codes for special purposes",
-    "26 Supplementary Chapter Traditional Medicine Conditions - Module I": None,
-    "V Supplementary section for functioning assessment": "Functioning assessment",
-}
-
-
-def find_roots(item, roots_lookup):
-    id = item["ID"]
-    if id in roots_lookup:
-        return roots_lookup[id]
-
-    if not item["parentId"]:
-        roots_lookup[id] = {item["classKind"]: item["label"]}
-        return roots_lookup[id]
-
-    if parent := roots_lookup.get(item["parentId"]):
-
-        def my(x):
-            return item["label"] if item["classKind"] == x else None
-
-        roots_lookup[id] = {
-            "chapter": parent.get("chapter") or my("chapter"),
-            "block": parent.get("block") or my("block"),
-            "category": parent.get("category") or my("category"),
-        }
-        return roots_lookup[id]
-    # The following code is never executed as the `icd11.json` file is
-    # pre-sorted and hence the parent is always present before the child.
-    print("Full-scan for", id, item["label"])
-    return find_roots(
-        [
-            icd11_object
-            for icd11_object in fetch_data()
-            if icd11_object["ID"] == item["parentId"]
-        ][0]
-    )
-
-
-def roots(item, roots_lookup):
-    roots = find_roots(item, roots_lookup)
-    mapped = ICD11_GROUP_LABEL_PRETTY.get(roots["chapter"], roots["chapter"])
-    result = {CLASS_KIND_DB_KEYS.get(k, k): v for k, v in roots.items()}
-    result["meta_chapter_short"] = mapped
-    result["meta_hidden"] = mapped is None
-    return result
+from care.facility.utils.icd.common_data import fetch_icd11_data, icd11_id_to_int, roots
 
 
 class Command(BaseCommand):
@@ -138,7 +31,7 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         print("Loading ICD11 diagnoses data to database...")
         try:
-            self.data = fetch_data()
+            self.data = fetch_icd11_data()
             ICD11Diagnosis.objects.bulk_create(
                 [
                     ICD11Diagnosis(
